+++ distributedcombigrid/src/sgpp/distributedcombigrid/mpi_fault_simulator/Sim_FT_comm_free.cpp
@@ -8,6 +8,8 @@
 #include "MPI-FT.h"
 #include REAL_MPI_INCLUDE
 #include <iostream>
+#include <boost/numeric/conversion/cast.hpp>
+
 
 int simft::Sim_FT_MPI_Comm_free(simft::Sim_FT_MPI_Comm *f_comm) {
   // MPI_Comm_free is a collective function according to the standard, so we use our dead processes
@@ -46,7 +48,7 @@ int simft::Sim_FT_MPI_Comm_free(simft::Sim_FT_MPI_Comm *f_comm) {
       (*f_comm)->NBC_Vector_Send = simft::NBC_to_Vector((*f_comm));
     }
   }
-  for (int i = 0; i < (*f_comm)->NBC_Vector_Send.size(); i++) {
+  for (int i = 0; i < boost::numeric_cast<int>((*f_comm)->NBC_Vector_Send.size()); i++) {
     std::cout << "NBCOUT:::" << (*f_comm)->NBC_Vector_Send[i] << "\n";
   }
 
@@ -114,7 +116,7 @@ int simft::Sim_FT_MPI_Comm_free2(simft::Sim_FT_MPI_Comm *f_comm) {  // does not 
       (*f_comm)->NBC_Vector_Send = simft::NBC_to_Vector((*f_comm));
     }
   }
-  for (int i = 0; i < (*f_comm)->NBC_Vector_Send.size(); i++) {
+  for (int i = 0; i < boost::numeric_cast<int>((*f_comm)->NBC_Vector_Send.size()); i++) {
     std::cout << "NBCOUT:::" << (*f_comm)->NBC_Vector_Send[i] << "\n";
   }
